Move all methods into commands folder
