+++ commands/contribute.py
@@ -1,12 +1,17 @@
-from command import Command
+from commands.command import Command
 from fbchat import Message
 from fbchat import Mention
 
 
-class good_video(Command):
+class contribute(Command):
 
     def run(self):
-        response_text = "@" + self.author.first_name + "\nGood Videos Here:\nhttps://www.youtube.com/watch?v=3CQyYE1rMJg&t=6s\nhttps://www.youtube.com/watch?v=oHg5SJYRHA0\nhttps://www.youtube.com/watch?v=c5daGZ96QGU"
+        response_text = """
+        @{} 
+        \nYou can contribute to Dentaku here: https://github.com/VikingsDev/Dentaku
+        \nFor more information, check out vikingsDev Bounties: https://vikingsdev.ca/bounties
+        \nHappy contributing!
+        """.format(self.author.first_name)
         mentions = [Mention(self.author_id, length=len(self.author.first_name) + 1)]
 
         self.client.send(
@@ -18,5 +23,5 @@ class good_video(Command):
     def define_documentation(self):
         self.documentation = {
             "parameters": "None",
-            "function": "OMG I JUST SAW YOU IN THIS VIDEO!1!"
+            "function": "Learn how to become a bounty hunter."
         }
