+++ commands/events.py
@@ -1,6 +1,5 @@
-import json
 from pyshorteners import Shortener
-from command import Command
+from commands.command import Command
 from fbchat import Message
 from fbchat import Mention
 import datetime
@@ -22,8 +21,8 @@ class events(Command):
         # The file token.pickle stores the user's access and refresh tokens, and is
         # created automatically when the authorization flow completes for the first
         # time.
-        if os.path.exists('token.pickle'):
-            with open('token.pickle', 'rb') as token:
+        if os.path.exists('../token.pickle'):
+            with open('../token.pickle', 'rb') as token:
                 creds = pickle.load(token)
         # If there are no (valid) credentials available, let the user log in.
         if not creds or not creds.valid:
@@ -31,10 +30,10 @@ class events(Command):
                 creds.refresh(Request())
             else:
                 flow = InstalledAppFlow.from_client_secrets_file(
-                    'credentials.json', SCOPES)
+                    '../credentials.json', SCOPES)
                 creds = flow.run_local_server(port=0)
             # Save the credentials for the next run
-            with open('token.pickle', 'wb') as token:
+            with open('../token.pickle', 'wb') as token:
                 pickle.dump(creds, token)
 
         return creds
