+++ commands/good_video.py
@@ -1,17 +1,12 @@
-from command import Command
+from commands.command import Command
 from fbchat import Message
 from fbchat import Mention
 
 
-class contribute(Command):
+class good_video(Command):
 
     def run(self):
-        response_text = """
-        @{} 
-        \nYou can contribute to Dentaku here: https://github.com/VikingsDev/Dentaku
-        \nFor more information, check out vikingsDev Bounties: https://vikingsdev.ca/bounties
-        \nHappy contributing!
-        """.format(self.author.first_name)
+        response_text = "@" + self.author.first_name + "\nGood Videos Here:\nhttps://www.youtube.com/watch?v=3CQyYE1rMJg&t=6s\nhttps://www.youtube.com/watch?v=oHg5SJYRHA0\nhttps://www.youtube.com/watch?v=c5daGZ96QGU"
         mentions = [Mention(self.author_id, length=len(self.author.first_name) + 1)]
 
         self.client.send(
@@ -23,5 +18,5 @@ class contribute(Command):
     def define_documentation(self):
         self.documentation = {
             "parameters": "None",
-            "function": "Learn how to become a bounty hunter."
+            "function": "OMG I JUST SAW YOU IN THIS VIDEO!1!"
         }
